refactor: move InjectRenderer to render package
